requirements.txt
zulip
zulip_bots
dateparser
python-osc
pillow
